backend/requirements.txt
django==5.0.2
weasyprint==60.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
pytest-django==4.8.0
pytest-html==4.1.1
django-filter==23.5
whitenoise==6.6.0
argon2-cffi==23.1.0
typing-extensions==4.9.0
djangorestframework==3.14.0
django-stubs==4.2.7
coverage==7.4.1
django-tailwind==3.8.0
pyyaml==6.0.1
django-structlog==7.1.0
structlog==24.1.0
